+++ src/main/java/de/braintags/io/vertx/util/lock/LockedExecution.java
@@ -1,4 +1,4 @@
-package de.braintags.io.vertx.util.asyncLock;
+package de.braintags.io.vertx.util.lock;
 
 @FunctionalInterface
 public interface LockedExecution {
